# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API doc of UniqueStudio Recruitment. For more API information, please see https://app.apifox.com/project/2985744",
        "title": "UniqueStudio Recruitment API",
        "contact": {
            "email": "wwbstar07@gmail.com"
        },
        "version": "0.1"
    },
    "host": "https://dev.back.recruitment2023.hustunique.com/",
    "paths": {
...
        "/comments": {
            "post": {
                "description": "create comment for applications, only can be created by member.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "create comment for application",
                "operationId": "create_comment.",
                "parameters": [
                    {
                        "description": "create comment opts",
                        "name": "pkg.CreateCommentOpts",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.CreateCommentOpts"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pkg.Comment"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
        "/comments/{cid}": {
            "delete": {
                "description": "delete comment of application, only can be deleted by comment's owner.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "delete comment of application",
                "operationId": "delete_comment.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "comment uid",
                        "name": "cid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    },
                    "400": {
                        "description": "code is not 0 and msg not empty",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "pkg.Comment": {
            "type": "object",
            "properties": {
                "application_id": {
                    "description": "manytoone",
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "evaluation": {
                    "$ref": "#/definitions/pkg.Evaluation"
                },
                "member_id": {
                    "description": "manytoone",
                    "type": "string"
                },
                "member_name": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "pkg.CreateCommentOpts": {
            "type": "object",
            "required": [
                "application_id"
            ],
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "evaluation": {
                    "$ref": "#/definitions/pkg.Evaluation"
                },
                "member_id": {
                    "type": "string"
                },
                "member_name": {
                    "type": "string"
                }
            }
        },
...
        "pkg.Evaluation": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "Good",
                "Normal",
                "Bad"
            ]
        },
...
    },
...
}